refactor(kkernel): expose the CLI as kkernel::cli::cli_main
Moves the clap command surface from main.rs into the library so a
downstream distribution can embed the full CLI in its own binary with
additional pack crates force-linked (ADR-027 self-registration). The
kkernel binary becomes a one-line shell; no behavior change.
